+++ Sources/RedBlackTreeModule/RedBlackTreeSet.swift
@@ -915,3 +915,4 @@ extension RedBlackTreeSet {
     other.forEach { insert($0) }
   }
 }
+