+++ Sources/RedBlackTreeModule/_RedBlackTree/___RedBlackTree+Pointer.swift
@@ -128,3 +128,5 @@ extension ___RedBlackTree.___Tree {
     }
   }
 }
+
+