+++ Tests/RedBlackTreeTests/MergeTests.swift
@@ -12,12 +12,20 @@ final class MergeTests: XCTestCase {
 
   func testSetAndSet() throws {
     var lhs: RedBlackTreeSet<Int> = [1, 2, 3]
-    var rhs: RedBlackTreeSet<Int> = [4, 5, 6]
+    let rhs: RedBlackTreeSet<Int> = [4, 5, 6]
     lhs.insert(contentsOf: rhs)
     XCTAssertEqual(lhs.map { $0 }, [1, 2, 3, 4, 5, 6])
     XCTAssertEqual(rhs.map { $0 }, [4, 5, 6])
   }
-  
+
+  func testSetAndSwiftSet() throws {
+    var lhs: RedBlackTreeSet<Int> = [1, 2, 3]
+    let rhs: Set<Int> = [4, 5, 6]
+    lhs.insert(contentsOf: rhs)
+    XCTAssertEqual(lhs.map { $0 }, [1, 2, 3, 4, 5, 6])
+    XCTAssertEqual(rhs.sorted().map { $0 }, [4, 5, 6])
+  }
+
   func testSetAndMultiset() throws {
     var lhs: RedBlackTreeSet<Int> = [1, 2, 3]
     let rhs: RedBlackTreeMultiset<Int> = [4, 4, 5, 5, 6, 6]